requirements.txt
# Page Classification System - Dependencies
# Python 3.10+

# HTTP and crawling
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
# Browser rendering (optional, for SPA support)
playwright>=1.40.0

# LLM integration
openai>=1.12.0

# Configuration
pyyaml>=6.0.0
pydantic>=2.6.0
pydantic-settings>=2.2.0

# Storage
aiosqlite>=0.19.0

# Utilities
tenacity>=8.2.0
python-dateutil>=2.8.0
